Implement MIDSX library and update VoxelGrid constructor for Python checks
Created new CMakeLists, MIDSXConfig.cmake.in, and hvl.cpp files for the implementation of MIDSX library. Updated CMakeLists.txt to build a MIDSX static library, locate and link Eigen3, and organise the installation and export of targets and files. Furthermore, VoxelGrid's constructor in voxel_grid.h and voxel_grid.cpp is updated to accept a new argument 'is_python_environment' to check if Python interpreter is already initiated. Additionally, removed 'computeBinWidth', 'computeCosinesSumForEnergyBin', and 'isWithinEnergyBin' methods in quantity_container.h/cpp for they were not being used anywhere.

These changes were necessary for the proper functioning and organisation of the code as well as to prevent redundancy. The update in the VoxelGrid's constructor is important to ensure that Python interpreter isn't started twice when working in a Python environment.