Add DiscIsotropicDirectionality to directionality
A new class, DiscIsotropicDirectionality has been added to the Directionality. The new class allows sampling directions uniformly over a disc with given center, normal vector and radius. The DiscIsotropicDirectionality class contains methods to set orthogonal vectors u and v based on the normal vector, and a method to calculate a vector along the circumference of the disc for a given angle.